DeleteIcon now properly removes data folder.
Icon images are still heavily volatile, and are/can be broken very easily.
Recommended to just rebuild your desktop from scratch in a few days when this whole process is completed.

DeleteIcon has been updated to use the id instead of row,col. And it fully deletes the icon + the data folder upon deletion.