editIcon, getDesktopIcon, defaultDesktopIcon updates.
EditIcon (ipc call, window, component etc.) now updated to receive the id as a parameter.

The temp id is assigned now in the DesktopGrid before calling EditIcon for new icons.

getDesktopIcon now takes an id instead of row,col since this is not safe as row,col can be different based on future profiles.

DefaultDesktopIcon now includes an id parameter as id is now a required field for a desktopIcon.

Editing an icon now shows editing[id] at [row,col] for the header for debug/easy identifying the id/row+col.