Beginning of DesktopProfiles. (Creates Desktop Profile folder).
Working towards an implementation of profiles for Desktop icons.

For now this just creates a appData/Roaming/AltDesktop/profiles/default/profile.json.

This for now just copies the saved desktopIcons.json.

Working on the overall design and for instance if desktopIcons.json should be kept in some capacity or just removed.