EnsureUniqueId now compares folder names for uniqueIDs, and fixed EditIcon renaming.
EnsureUniqueId now compares folder names in /data/ to get a unique ID. This is a better solution to ensuring that ids are unique and the folders do not get merged together accidentally depending on the future implementation of desktopIcons.json and profiles.

Also fixed a mistake where it was accidentally sanitizing the userID in EditIcon thus would always rename the base name folder when it was opened (temp if new) to the newId name.
An example of this, is if you had an icon named temp, and created a new icon, it would always take the temp image folder, and rename it to the newId, thus breaking both the temp icon image, and the new icon image. This now uses the proper ids so it should avoid such issue.