/profiles/default/profile.json Now Default profile. And fixed EditIcon for id/names with spaces.
The program now uses this /profiles/default/profile.json instead of "desktopIcons.json".
If this does not exist, it will copy from desktopIcons.json, if you have a current setup.
desktopIcons.json still exists, not sure if it will be used. I might turn it into an index of icon ids, where if an icon is in that file but not in a new profile it knows to add it to a new desktopIcon in that profile. For example you add an icon in one profile, then swapping profiles it shows it at the earliest free space, then you can move it from there, or hide it.

Fixed the subWindowHash for EditIcon's id parameter to encode the id so it can properly load icons with spaces in the name/id.