Map/Index redesign for DesktopGrid.
Now stores the primary map of icons (iconsById) as a map of ID -> Icon.
Also retains a posIndex which stores posKey -> icon id.
(Still retains getIcon(row, col) which gets the posIndex id then gets the id from iconsById map).

Keeps the posIndex and the iconsById in sync.

Going forward this seems like the better approach as it it allows direct calls from say reloadIcon ipc event to instead call with the id and it will reload it no matter where its row,col is.
For example if you edit the row, col in profile.json for right now, then try reloading the icon, it will make the icon disappear but not reload the actual icon, thus you just deleted it and must reload the desktop to make it appear. This will be fixed to instead listen for an ID to reload instead of the row,col which are possibly changed.