ipc call reloadIcon(row,col) -> reloadIcon(id)
ipc reloadIcon() no longer relies on the row,col to reload the icon. It now just sends a message to reload an ID which it then fetches and reloads.

This ensures that even if the row/col has changed it will still reload the icon. For instance previously since it required the row,col it would only reload the icon at row,col and if there is no longer an icon there, it would just remove it and not render the new row,col.