removed limitation on interacting with desktop while subMenu opened
Removed the limitation always nulling right clicks when a submenu is open.
Originally when implemented I wanted 100% to make sure the user does not lose the subMenu window, and wanted to ensure that the user does not accidentally open a new subWindow without saving changes etc. so I locked it down.

However in the background I set up the system correctly for what I wanted, can't open a new subWindow while one exists, attempting to will re-center the subWindow on the current screen. A fix for both menus that go off screen and avoids problems like closing without saving.