Icon generation, combined apporach.
Reported false positive for the nuitka to Microsoft. They responded: "At this time, the submitted files do not meet our criteria for malware or potentially unwanted applications. The detection has been removed. Please follow the steps below to clear cached detections and obtain the latest malware definitions."

So they seem to have unmarked it as malicious.
Unfortunately the problems with some executables which the rust file_to_image has doesn't seem easily fixable. It continues to generate some specific executables with icons super tiny in a padded box. The problem is even windows does this for example if you look up the executables which do this in windows search they appear in the padded box. I'm not 100% how to fix it.

For now I am re-including the python file_to_image (now named exe_to_image) which removed the default program methods and now just fetches the icons from executables.
This icoextract somehow doesn't fall into the pitfall of the rust file_icon_provider, which returns the padded icon file.
Once again technically this is a rare occurrence and I will try to work out a way to fix it in rust to just have 1 standard fast solution.