getDesktopIconData() now takes profile name.
getDesktopIconData now takes a profile name, and attempts to get the profile.json for that folder. For instance "cool_background" profile would look like this: "AltDesktop/profiles/cool_background/profile.json".

Does not yet work with saving the profile, all updates like showing it are rendered locally and instead save into the /default/profile.json at the moment. Needs a bit of work before the profiles are stable, which is why I am not implementing a way to change profiles.

Funnily enough, I thought it would just be rendered locally via preview but if you save, and reload the icon it accurately reloads the icon since the reload icon also fetches from the default profile.json....