ID syncing across profiles.
One flaw in the idea for the profile structure was that renaming an icon changes the id. This can be kept in sync by updating all ids in the profile.json files to change. Not my happiest design, but should work.

Icon ids are guaranteed to be unique unless copied due to how it selects new ids based on the folders stored in /data/. The only real issue with this is that in the file system the id of the folders will be the last time you renamed an icons name, not necessarily the name of the icon (unless you only use default profile). The good thing about this though is that you can always open up the icon folder directly via right click -> open -> image folder.

Major note: editIcon is still saving only on default profile. Profiles are still unsafe, which is why no options to swap it yet. Also fixed the reload-background useEffect to not change profile by default (accidentally left the code to swap profiles in testing in the actual commit).