profile is now a RendererState.
Profile is now a rendererState and updates all components whenever it is changed. It also defaults to "default" now instead of just "". Will likely use the "" state for debugging, and avoiding accidentally corrupting saves/changing default profile when profile is not passed as intended.

Reminder line 933 is the line to test changing to different profiles for now.