getRendererState simplification.
Added a getRendererState to rendererStates in main process which just returns one, key's value which is called.
Example: await getRendererState("profile") returns active "profile" value.

Then simplified the logic for use in ipc functions such that it fetches it as a profile, if it does not return or if it returns "" it shows a warning. Then if it exists it just const creates the filepath to read/write to, instead of creating a template, then returning due to error.