Added profile in BackgroundSummary local data.
Added a "local", "profile" option to bg.json files.
By default it saves with "default". No option to change to another profile yet.
Also fixed a case where passing with just an empty BackgroundSummary ("{}") was not creating the initial summary as expected, due to even with a value of "{}" it was not returning false as it existed. It now checks to see if the keys > 0 if not, then creates the new id.