ensureProfileFolder("profileName") when loading/reloading background.
Now ensures a profile, and the profile.json exist for the saved profile
when loading/reloading a background.

example: create/edit a background to have a new unique profile name, save it, close backgroundSelect, it will reload and create the /profiles/"profileName"/profiles.json folder (initially empty). This is now an active working profile and you can add to it. This currently works as a "soft" creation for a profile, i.e. it ensures that a profile always exists.
In the future creating a new profile will have options like copying from existing.