LaunchProgram with no profile now assumes "default" profile.
Updated launchProgram to default to "default" profile if it does not detect a profile in the selectedBg.

Also removed the smallWindow warning for getDesktopIconData, as this will appear even when no backgrounds exist (first launch).

I think the idea is to just make !profile assume its using "default" in all areas, and treat it as such.