Keep selection on exiting EditBackground.
Not happy with the current setup, but fixing 1 part of a two point bug.

Reloading this background essentially goes back to the saved value previewed, despite it still updating the selectedBg to the bg that was open and selected.

Not happy with the whole backend convoluted way that backgroundSelect attempts to load and display everything. It has multiple race conditions that consistently fail to work properly. Will start to work on this.