closing EditBackground without saving now re-selects selected bg.
Updated selecting a background to now call to save settings data to keep track of your last selected background.

This fixes a problem where it would reload with the settings saved background when re-opening backgroundSelect via: window.electron.openBackgroundSelect();.

I had intially made it such that we would just pass the summary.id to re-select but that does not guarantee it was the last selected background as a user could select one background, then right click edit another without ever selecting it, so by having (summary.id) it would re-open with the summary.id which was not the selected background previously. So by calling reloadBackground() and openBackgroundSelect() we can effectively reload the background (to drop the preview), and open it which will default it to the last manually selected background.