100ms delay before showing BackgroundSelect page when opening/changing page.
Added a 100ms delay, where it shows a loading-overlay (just a same color as the background box overlay over the grid of backgrounds).

This helps prevent the images visibly loading (where it loads the top first, down to the bottom like old dial-up days).

This will not 100% ensure that all images are loaded before showing, but the 100ms gives it a good headstart which should work for most computers. Some slow pcs might need a bit longer, but it should keep it to just a small delay that works for most without resorting to too long of a delay for fast computers, or delaying it strictly until all icons are loaded. A good mid range usually works, and barely noticeable approach is best.