Now cleans up temp_# folders when not saving new icons.
Now when closing without saving for new icons it calls to delete the icon data folder as well, to cleanup the temp folders for image generation (which follow the pattern of temp, temp_1, temp_2 etc.).

Reminder, users can still have icons with the name/id "temp" or "temp_1" etc. But these temp folders receive unique "temp_#" folders which will call to delete themselves if the user decides not to save (which when saving will update the id/name of the folder).

Still need to implement a way to actually delete the /data/ folder if it is only used by a single profile. Will work on that.