AddProfile module EditBackground (work in progress).
Added a AddProfile module for EditBackground which works like the modules for AddTag and EditCategories. For now it just creates the small module which allows the user to type in a name to create a profile.

Does not yet add a profile, just logs "Attempted to make profile: profile_name". (Needs backend support for it to fully work) Then will require more work for ability to copy from existing backgrounds and further work to allow copying specific icons etc. from other profiles. As well as when creating a profile does this profile auto add it to this profile etc. Lots of work to do.