Added visible warning for adding profile which already exists.
Added an inline error to warn the user that the profile they are trying to create already exists.

Moved away from the situation where its fine just because it just ensures that it exists. For example do not want a user to create a profile called "cool profile" when they already have a "cool profile" thus they think they added one but in reality it already exists, then they go to edit it not realizing they are editing the existing profile as well.