ensureProfileFolder now has option to copy from.
Now when calling ensureProfileFolder, it takes an extra optional parameter "copyFromProfile?"

When called without copyFromProfile:
ensures profile folder, and profile.json (blank) are created.

When called with copyFromProfile: it will always replace the profile.json (if it exists, or will create it from scratch) to match the profile string passed.

For example:
await window.electron.ensureProfileFolder("new_profile", "default");

Will ensure that new_profile folder exists, and that it has a profile.json which DIRECTLY matches the "default" profiles profile.json.
If the profile.json already exists for new_profile and this is still called, it will renamed the existing profile.json to "backup.json" to hopefully catch and allow a revert if this is improperly called somehow.

Shows a warning if the profile.json for the profile to copy from is not found (and does not change the profile.json).