deleteIcon ipc now deletes /data/ folder if icon is unique to profile.
Now when deleting an icon, it checks the other profiles for an icon with the same id.
If found, it returns as normal (to retain the /data/ folder as other profiles are using it).

If not found, the icon is unique and it will call to delete the /data/ folder (it moves it to the recycle bin, does not directly delete).