Drag + drop moving desktopIcons. (early in development).
Can now drag and drop icons into different locations.
Dragging creates a copy of the dragged icon which follows the cursor.

Dropping the icon will move the icon to the hovered location.
Does NOT allow dropping on existing icons, for now it just shows an error and reverts the drag.

Will implement a better way for previewing where the icon will appear such as highlighting the box it appears in. I will probably make the default "snap-to" the icon box (0 x/y offset). But add in a keybind to allow placing exactly based on mouse location so you can adjust icons offsets without manually typing in values.