moveDesktopIcon ipc added param: offsetReset?: boolean .
Added an optional offsetReset: boolean parameter to moveDesktopIcon.

Purpose for this is an optional way to reset the offsets when moving the icon.

Current plans is by default just dragging it will reset the offsets and simply place the new icon at the icon home base location.
Then will add a ctrl modifier which will allow the user to place the icon freely, and will create its own offset based on where the user places the icon.
Lastly might add a way to just move and maintain the offset, but unlikely.

For now only implemented in ipc call:
await window.electron.moveDesktopIcon("icon_id", 0, 0, true);
Where this would move the icon with the id "icon_id" to 0, 0 and reset its offset to 0 for both x/y.