Dragging icons now resets offsets by default.
As seen in last post where I added the ipc support for resetting offsets. It now resets offsets when drag/drop moving icons.

i.e. the previewed icon when dragging is exactly how it will appear when you drop it.

Will work on adding the ctrl modifier which allows the user to place it with offsets based on their mouse.