Fixed non-unique icons updating their ID every save.
Fixed an error where it would compare the current icon name to the icon id.
Where if the id of the icon was say "icon_1" and the name was "icon" it would believe the name was changed, and thus call to get a uniqueIconId which would give it the id: "icon_2".
Therefore it was possible to rotate between "icon_1", and "icon_2" every time you saved the icon.