Dropping on an existing icon now swaps icons.
Dropping a dragged icon on an existing icon now swaps icons.

Note: This does reset offsets for both icons. The reasoning for this is that if you are ever swapping icons and not ctrl modifying to place it precisely (not yet implemented). You are essentially snapping it to grid in new location. Also swapping it and not resetting an offset could be confusing such as swapping an icon and it now appearing offscreen based on where it was swapped to.

Further when ctrl modifer does get implemented it will try to essentially find the nearest empty icon home (prioritizing left and above) then setting the offset to put it where it shows the user is placing it.

So the current idea is
normal dragging: snap icons to grid (remove offset).
ctrl dragging: place icon where user places it by mouse, update offset to keep it there.