setRendererStates now checks profiles to ensure they exist.
Now when setting the renderState for a profile (previewing or loading a background), it performs a check that the profile exists. If it does not exist it creates it.

Previously it would just show as blank and then had issues actually saving with "default" when editing the background. (another thing to fix).