Fix some bugs in EditBackground
Fixed previewBackgroundUpdate still using the old bgFile path instead of the id it expects. (Not sure how this was still working to be honest).

Background Paths that do not exist or are not set no longer recommend "Portrait" global tag (or "other").

Attempting to save with a background that does not exists no longer gives the error invalid type "[]". Instead gives the error Background File not found.

Still some other things to work out, but fixed the easy immediate issues.