Misc. tweaks. (Profiles, mostly safe from now on).
Removed warning that profiles/desktop icons were volatile. I think profiles are mostly safe at the moment. The only real thing it is missing is a menu where you can change profiles (without needing to go through editBackground), remove/rename profiles, copy from profiles, and even view profiles icons and import to current profile. It will be a bit of work, but the underlying format for profiles should not change, just needs some functions to read/copy from profiles to different profiles, so for the most part it should be safe for the foreseeable future.

Some random tweaks I found messing around in some menus.

Now shows a smallWindow message that it failed to get bg.json if it fails to read. Also shows the error so can help user understand. Not the most verbose of errors especially for improper json format for example, but can help user understand why things don't look as expected.

When tweaking the size the icons must be before adding a new row, with larger icon sizes (240+) It became noticeable that it would add in 6 pixels when selecting an icon to a row, thus moving the row above up, the row below down etc. due to the border. This has been fixed to use the borders size in padding, then removes the border on selecting (thus maintains the same size).

Also removed the max-width of icons. Originally this was just set to a decently high quality, but since it fetches the original image regardless no performance hits for removing it. This was primarily removed because at a certain width/height it would stop growing but the icon box would continue to grow horizontally leading to just empty space. So it has been removed such that icons can continue to grow up until theres enough space for another row.