Changed page size to 60.
Sat there for a while trying to get a page size which would work well for a range of icon sizes without constantly having last rows which only contained a few icons instead of a complete row.

While 50 is nice and logical for a user to guess the total icons on a page if they were say trying to guestimate their total icons by page x icons per page. Having a page size of 60 is extremely good visually as all icons per row < 7 fit exactly. Leaving only 7, 8, 9 which do not fit exactly under 10 icons.
General expected use of this is 3-6 icons per row, so this covers a majority of the usecase.