Changing icon sizes in BgSelect + contextMenu tweaks.
Can now change sizes of icons in BackgroundSelect.
Note: this is only for the current open window, closing, or re-opening will reset back to small icons (just need to make this a setting).

Tweaked submenu's css to be more compact (smaller font + less padding for each field), and moved it to App.css as it has not been unique to DesktopGrid for a while.