Added bgSelectIconSize to settings.json file.
Added it as an option with a default value of "medium". However the BackgroundSelect window does not load from it or save yet.
This is just saving it as a new setting on startup.

Also re-organized the SettingsData object, this will not re-organize an existing file, but keeps the settingsData more readable for newer files by putting all the Settings Window options first, then followed by the hidden settings. (Things which the program manage and do not need the user to manually change.)