Fix one part of the 2 icons next to each other dragging.
Fixed the first issue where dragging an icon on top of an icon next to it.
It still has the drag over -> drag back -> drag back over issue where it renders the dragged over icon over the dragged icon, it works so long as you move to a different icon home, but when you move back and forth it does this constant cycle.