Fixed swapping icons.
Fixed a problem in the rewrite where icons are now directly previewed instead of the mock icons.

Problem was that getRowColFromXY would return based on the previews, so when swapping two icons it would end up as dropping it on itself since it was being previewed in the drop location, not to determine if it is getting swapped it just sets existingIcon to swapPreview?.icon which is nulled if user drags over a non-existing icon, which then it just tries to move it (not swapping).

Still having a lot of trouble getting the bug where two icons going back and forth can cause the 2nd icon to reload over the preview of the dragged icon instead of just reverting back and swapping.