Reverting back to commit: dd2c2eb (Temporary icons are back).
Reverting back to before the commit: Partial simplification of previewing dragged icons/dragged over icons. (16116d1)

Spent hours working to try to get back caught up with this original version while using handlePreviewUpdate(). Made good progress but I ran into the brick wall.

Basically the way that React acts when calling handlePreviewUpdate() is that it causes a re-render which nullifies the current drag (not visually, visually it still worked fine, but instead: onDragEnd={handleDragEnd}).
For reference as soon as it causes a handlePreviewUpdate() it will never call handleDragEnd. Thus when trying to Escape out of the drag, it just stops updating the drag (doesn't drop, but just stops updating). Which leaves the preview on screen and does not reload the icons.

The only real solution is to just revert back to the temporary icons which don't trigger full re-renders which break the drag. I thought I was  simplifying it, but it lead to such a huge mess.

Will look over this setup again, try and simplify some stuff, and add back the dimmed icon at the origin I had working in the last commit.