Fixed closing without saving icon previewing in next Edit.
Fixed a situation where closing an new icon without saving, and then attempting to make a new icon in the same spot, would have the old preview information stored, and if you changed any field it would only update that field, but re-preview the old unsaved icon.

Technically this does not fix that an icon does not get fully removed from the Map (so any unsaved icons technically are still in RAM until you edit the same icon and it previews it as a default blank icon.