Properly remove unsaved icons after closing EditIcon.
I guess as a failsafe I had left the references in the iconsById map, but only removed the posMap positions. This now properly removes the iconData from both maps. Meaning that if you close without saving an icon, it should properly remove it from RAM. (And just in general keeps the iconsById map clean as it will only have existing icons).