Now verifies read/write permissions on saving ExternalPaths.
Now when saving externalPaths, it will ensure the directory exists, and it can read/write to the directory before saving.
Now shows smallWindow errors for if the directory does not exist, or it cannot be written to.

I would add a check to better ensure that the user properly inputs them (file paths separated by ","). But in the future I plan to expand the basic way you input external paths rather than just typing them with a comma in between.