defaultBackgroundPath added to settings.json.
Not fully implemented as there is much to handle with it. But you can edit the Settings.json and change the defaultBackground path to something like "E:\\AltDesktop" and it will work (so long as you can read/write to this file, as I havent added a check for this yet). Keep in mind this will not transfer your existing default save location backgrounds yet. It is also heavily in testing and has a few bugs to try and fix.