Fix errors on moving Background to other Paths.
Fixed Permission error being shown when it moves, but fails to delete the old file (due to still being in use).

No longer errors and instead saves, then attempts to remove the folder after a delay (starts at 5 seconds, re-attempts 5 times over a few minutes.)

This ensures that the files are moved, and tries to ensure that the old files are cleaned up properly so as to avoid moving it back and forth causing it to create background_1, background_2, background_3, and just keep the static id it was given.