Changing Default Background Folder now asks how to import by date.
Changing the default background folder, if backgrounds are detected (folders with bg.jsons) it will prompt the user how they would like to import these backgrounds. Either as new or with their last indexed date.
This works exactly like External Paths  (now External Background Folders).

NOTE: Does not yet ensure that the Default Background Folder is read/writable, currently if it is not then it will cause an error on opening BackgroundSelect. If it is read/writable but does not exist it will cause the same error and upon restarting will create the folder. Further it currently does not have options for transferring from the existing background folder to the newly set one (Might add option for this later, but also might just rely on the user moving the backgrounds first then adding the default path and importing with saved date) <- very easy and hard to screw up.