Optimize getBackgroundSummaries.
Since we are already grabbing the bgJson. We have the icon and the bgFile arguments, instead of calling idToIconPath/idToBackgroundPath which reads directly from the file to obtain, we can just manually add the folder path to the icon/bgfile to get the full paths.

This saves us 2 reads per background on the page in BackgroundSelect.
With 60 per page, this saves 120 reads per page loaded.