Strengthened EditIcon box highlight in DesktopGrid.
Now when Edit Icon is opened for a new icon/Exiting icon it further solidifies that that icon will be highlighted for the duration of editing.

I had previously made it such that when right clicking on the DesktopGrid while a window was open would not go through, this is no longer the case and you can still right click the desktop while editing/in any other menu. So the problem was the user could edit an icon, then right click any where and it would remove the highlight from the editing icon. This now has evolved to use its own set highlight with pulsing, and only gets removed when a subwindow is closed.