Fix Non named icons swapping _1 -> _2 -> _1 -> _2 back and forth on save.
Fixed a scenario where blank unset names were failing a check, thus calling to set their ID to a new uniqueId with the base unknownIcon.

This would endlessly swap back and forth between say: unknownIcon_1 and unknownIcon_2. Because when saving it would say "hey i'm a new icon, let me get a unique name for a new icon" so it generates a unique ID which is unknownIcon_2. Then when opened again and saved it would say it again but this time _1 is free so it would become _1 etc.

Solved this by detecting if the icon is actually new, if it is then it gets an unknownIcon id (if left blank for name). Otherwise it uses the id it loads from when EditIcon is opened.

Also updated the npm modules (Last night's alpha 0.0.1 was compiled with this updated npm modules.)