Remove visual "default::" in backgroundSelect for default path saved backgrounds.
When using a manually defined in settings: Default Background Folder. The backgrounds in the root AppData/Roaming/AltDesktop/backgrounds/ will still show.
When not saved with a name they default to just showing their ID.
(This has now been changed to not include the default:: prefix").

This was a purely visual change as it just removes the prefix from the label. The background still correctly loads, and all other references like opening folder, deleting etc. still use the full id.