Fixed previews maintaining fields like offset and dimensions.
Fixed a problem where old IDs can remain in the iconsById map in DesktopGrid after saving a new/existing icon in EditIcon.

Now when saving an icon in EditIcon it calls for a reload of the original ID if the id has changed. This ensures that handleIpcReloadIcon() in DesktopGrid completely removes the old ID from the iconsById map.

Without this the icon could still remain and since the same "temp" icon ID is set every time you open a new icon, it would maintain the old offsets and size dimensions of the last saved icon, and improperly preview the new icon.