Slightly increased subWindow delay, decreased BgSelect delay.
Showing a subwindow now takes 200ms (0.2 seconds). This helps ensure the window properly renders properly before showing to the user. This helps avoid stuff like items shifting, images still loading etc. but keeps it short enough the user would not recognize it.

Decreased the 200ms delay on BackgroundSelect to 100ms to allow it more time to load for the first time, without copying that assigned time to swapping pages as well.