Detecting and moving AppData "Default" background folders on save.
Fixed a situation where when the user sets a "Default Background Folder" in settings, on attempting to save a background which still exists in the AppData folder, it would not attempt to move it to the new Default Background Folder specified in settings.

Basically the problem is that it was not recognizing it as needing to be moved because it was not detecting the default:: id prefix as not being a normal id.