Fix moving selected Bg to another folder unselecting Bg.
When you edit a background which was selected in BackgroundSelect, it now will update the id to the new folder reference if the user changes the "Save to". This maintains the selectedBg in its new folder when returning to BackgroundSelect.

Previously it would not properly select the background with the changed id, and would only select it if the user clicked "Save & Apply".