Require Absolute Paths for canReadWriteDir checks.
Was not aware that canRedWriteDir would return true when called with an incomplete path like E:AltDesktop (Should be E:\AltDesktop as it is missing the folder "\" signifier.)

This would cause weird problems as it would correctly save/move backgrounds to the directory but certain things like loading from it would not work due to how the files were accessed.

So both Default Background Folder, and External Background Folders settings require absolute paths.