Added minWidth/height to Settings window and ensuring it fits on screen.
Added the height calculations from other windows to this to ensure that if the default height/width are too large to fit on screen it downsizes it.

Also added a minWidth/Height as a minimum size viable for effectively using the window. This size is quite small so should work well regardless of the display a user is using as it is only 465x300 pixels unlike the backgroundSelect and other windows which have min values in the 500-600s