Simplify detectOffscreenIcons to calc once, and compare .row/.col.
Simplified it to instead calculate the max row and max columns that would fit in the area. Then checks each icon if it is within those max row/col. If not it is considered offscreen.

Multiple other small things where I am trying to find out why dropping ANY icon (even non offscreen panel) now makes the offscreen panel disappear. As far as I can tell it is happening due to a stale reference to default icon size/paddings Thus with setting for 74pixel icons, it correctly tags them as offscreen, but when detectOffscreenIcons is called again it seems to be reverting back to 64pixel icons which would make those icons NOT offscreen, with the 64 being a stale default reference to the icon size.