offscreenIconsPanel now shows by default on launch.
offscreenIconsPanel now shows by default on launch.

For now this means it will always show, I will introduce a "x" or hide soon which will hide it. But it will always appear on launch to quickly inform the user that some of the icons in the profile are offscreen. (to prevent users from accidentally pushing icons off screen then wondering where they went).