Can now hide Off-screen Icons panel (the one that appears by default).
When the panel shows on loading a profile/launching the window, it now has a "X" button which you can hide it.

However, when it appears in the Display -> All Icon Boxes, it cannot be hidden except for by disabling the All Icon Boxes.

This allows the program to tell the user when opening/swapping backgrounds that some icons appear off screen, but also allow them to keep them off screen if they wish. While still allowing them to show all icon boxes as a sort of debug menu for why some icons might not be on screen.