Resizing now detects offscreen icons.
Previously it was only setup such that the user changing the size of icons/default icon size would update and detect icons which are now offscreen/now onscreen.

It now keeps track of the viewport size, and updates when resized to re-run detectOffscreenicons.

So un-maximizing it, will update offscreen icons the same as resizing it manually.
The viewport useState was added with a 200ms timeout to avoid constantly updating while a user spam drags back and forth.
Spam dragging back and forth will always create some problems, but with the timeout it is about 50% more cpu efficient (While spam resizing: 11% cpu usage vs 17% without the timeout)