Icon Overview (Replacing Display -> All Icon boxes /show grid).
Combined All Icon Boxes and Show Grid into one "Icons Overview" in the right click menu and placed it at the top of the visual settings in the contextMenu. The improved accessibility should help users easily toggle it, and help new users discover it, as opposed to the hidden submenu previously.