detectOffscreen now detects offsetX/Y offscreen icons.
For icons which are offset (have a offsetX or offsetY !== 0) it now checks and ensures that even with the offsets, it is still on screen, if it appears offscreen then it adds it to the offscreen icons panel, even if its iconBox is still on screen, but the icon isn't.
For icons without offsets (offsets = 0) it has a miniscule performance increase by just checking the row/col.