When in EditIcon, Icons are no longer draggable.
Now when editing an icon it locks icons such that they cannot be dragged to move/swap etc.

I believe 99% of it was technically fine, so long as you didn't move the icon you were editing, but in order to block an absolute massive amount of possible bugs its best to just lock them.