Simplify EditIcon to just preview entire icon.
Instead of ensuring that every single option in EditIcon both sets the icon, and sends the preview with that fields update, it now just sets the icon. And when Icon is updated it sends the preview of the entire icon. Much simpler and more resistant to accidental bugs like not sending previews.