editIconOffsetUpdate ipc command.
Added an ipc command for updating the EditIcon's offsets. The basic idea behind this is that the icon being edited will be draggable, allowing for the user to just drag the icon to the location they want and it will update the offsets to fit it there.

Call via:
await window.electron.editIconOffsetUpdate(10, 100);
Where this sets the offsetX to 10, and offsetY to 100.
Note: this requires an Edit Icon subwindow to be open to actually work, if the window is not open it does not work.