Dragging Edit Icon to reposition.
Can now drag an Icon while Editing it on the desktop. Doing so will update the offsets which keeps the Edit Icon window update such that it will save correctly.
I think I can still optimize this a bit as it is not the best performance due to multiple ipc calls/sends every pixel moved.