Optimized EditIcon DeskopIcon dragging.
Now while dragging the icon it just updates the offsets locally for the icon. Then upon dropping updates the EditIcon Icon which then just does 1 quick preview.

Previously for every single pixel it was moved it would do the following:
 ipc call (update offset) -> webcontents.send (to EditIcon) -> receiving -> updating states -> ipc call (preview) -> webcontents.send (to DesktopGrid) -> receiving -> updating icon state.

Now it only does this once on dropping the icon, and simply updates the Icon offsets locally to preview it before dropping it.
About 50-60% more efficient for cpu usage.

Note: Dragging these icons while editing do have the base electron dragging system of Esc key "dropping" the icon where it does not actually call the drop function.
In the future this will be changed to instead stop dragging and reset the offsets to the offsets the icon had when opened for the first time.